+++ test/main.test.js
@@ -30,6 +30,80 @@ describe("main.js", function () {
 
       unix.escapeShellArg.restore();
     });
+
+    it("works for boolean values on 'win32'", function () {
+      let outputTrue, outputFalse;
+      try {
+        outputTrue = escapeShellArgByPlatform(true, "win32");
+        outputFalse = escapeShellArgByPlatform(false, "win32");
+      } catch (_) {
+        assert(false, "Should not throw for a boolean");
+      }
+
+      assert.strictEqual(outputTrue, "true");
+      assert.strictEqual(outputFalse, "false");
+    });
+
+    it("works for boolean values on 'linux'", function () {
+      let outputTrue, outputFalse;
+      try {
+        outputTrue = escapeShellArgByPlatform(true, "linux");
+        outputFalse = escapeShellArgByPlatform(false, "linux");
+      } catch (_) {
+        assert(false, "Should not throw for a boolean");
+      }
+
+      assert.strictEqual(outputTrue, "true");
+      assert.strictEqual(outputFalse, "false");
+    });
+
+    it("works for number values on 'win32'", function () {
+      let output;
+      try {
+        output = escapeShellArgByPlatform(42, "win32");
+      } catch (_) {
+        assert(false, "Should not throw for a number");
+      }
+
+      assert.strictEqual(output, "42");
+    });
+
+    it("works for number values on 'linux'", function () {
+      let output;
+      try {
+        output = escapeShellArgByPlatform(42, "linux");
+      } catch (_) {
+        assert(false, "Should not throw for a number");
+      }
+
+      assert.strictEqual(output, "42");
+    });
+
+    it("fails for undefined values on 'win32'", function () {
+      for (const val of [undefined, null]) {
+        let threw = false;
+        try {
+          escapeShellArgByPlatform(val, "win32");
+        } catch (_) {
+          threw = true;
+        } finally {
+          assert(threw, `Should throw on '${val}'`);
+        }
+      }
+    });
+
+    it("fails for undefined values on 'linux'", function () {
+      for (const val of [undefined, null]) {
+        let threw = false;
+        try {
+          escapeShellArgByPlatform(val, "linux");
+        } catch (_) {
+          threw = true;
+        } finally {
+          assert(threw, `Should throw on '${val}'`);
+        }
+      }
+    });
   });
 
   describe("quote", function () {
@@ -65,6 +139,80 @@ describe("main.js", function () {
       const output = quoteByPlatform("Hello world!", "linux");
       assert(output.startsWith("'"));
       assert(output.endsWith("'"));
+    });
+
+    it("works for boolean values on 'win32'", function () {
+      let outputTrue, outputFalse;
+      try {
+        outputTrue = quoteByPlatform(true, "win32");
+        outputFalse = quoteByPlatform(false, "win32");
+      } catch (_) {
+        assert(false, "Should not throw for a boolean");
+      }
+
+      assert.strictEqual(outputTrue, '"true"');
+      assert.strictEqual(outputFalse, '"false"');
+    });
+
+    it("works for boolean values on 'linux'", function () {
+      let outputTrue, outputFalse;
+      try {
+        outputTrue = quoteByPlatform(true, "linux");
+        outputFalse = quoteByPlatform(false, "linux");
+      } catch (_) {
+        assert(false, "Should not throw for a boolean");
+      }
+
+      assert.strictEqual(outputTrue, "'true'");
+      assert.strictEqual(outputFalse, "'false'");
+    });
+
+    it("works for number values on 'win32'", function () {
+      let output;
+      try {
+        output = quoteByPlatform(42, "win32");
+      } catch (_) {
+        assert(false, "Should not throw for a number");
+      }
+
+      assert.strictEqual(output, '"42"');
+    });
+
+    it("works for number values on 'linux'", function () {
+      let output;
+      try {
+        output = quoteByPlatform(42, "linux");
+      } catch (_) {
+        assert(false, "Should not throw for a number");
+      }
+
+      assert.strictEqual(output, "'42'");
+    });
+
+    it("fails for undefined values on 'win32'", function () {
+      for (const val of [undefined, null]) {
+        let threw = false;
+        try {
+          quoteByPlatform(val, "win32");
+        } catch (_) {
+          threw = true;
+        } finally {
+          assert(threw, `Should throw on '${val}'`);
+        }
+      }
+    });
+
+    it("fails for undefined values on 'linux'", function () {
+      for (const val of [undefined, null]) {
+        let threw = false;
+        try {
+          quoteByPlatform(val, "linux");
+        } catch (_) {
+          threw = true;
+        } finally {
+          assert(threw, `Should throw on '${val}'`);
+        }
+      }
     });
   });
 });